moved modelbuilder to own file
